fix UI issue with newly added identity looking wrong in the UI due to using events instead of blocking
