+++ DesktopEdge/Models/UIModel.cs
@@ -1,14 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
+﻿
 namespace ZitiDesktopEdge.Models
 {
-    static class UIModel
-    {
-        static public bool HideOnLostFocus { get; set; } = true;
-
-    }
+    static class UIModel {
+		static public bool HideOnLostFocus { get; set; } = true;
+	}
 }
